host/describe_and_generate/requirements.txt
gradio
python-dotenv
pillow
